Semana 2/scrip de pyhton para las cloud function/requirements.txt
google-cloud-storage==2.7.0
pandas==1.4.4
numpy==1.23.2
cloudstorage==0.11.0
gcsfs==2022.11.0